Normalización endpoint testimonials
